Add error pages and customize blade layouts

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,12 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate(10);
+        //  ->where('user_id', request()->user()->id) to show only the notes for the user
+
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy('created_at', 'desc')
+            ->paginate(50);
         //dd($notes);
         return view('note.index', ['notes' => $notes]);
     }
@@ -34,7 +39,7 @@ class NoteController extends Controller
             'note' => ['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
         return to_route('note.show', $note)->with('message', 'Note was created.');
@@ -45,6 +50,11 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        // resticted access - only user who owns the note has access
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         return view('note.show', ['note' => $note]);
     }
 
@@ -53,6 +63,11 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        // resticted access - only user who owns the note has access
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         return view('note.edit', ['note' => $note]);
     }
 
@@ -61,6 +76,11 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        // resticted access - only user who owns the note has access
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'note' => ['required', 'string']
         ]);
@@ -75,6 +95,11 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        // resticted access - only user who owns the note has access
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note ID: ' . $note->id . ' deleted.');
